requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyScss==1.4.0
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
wheel==0.44.0
